+++ features/form/errorMessagesHandler.ts
@@ -1,4 +1,30 @@
-import { VaultErrorMessage } from '../openMultiplyVault/openMultiplyVaultValidations'
+export type VaultErrorMessage =
+  | 'depositAmountExceedsCollateralBalance'
+  | 'depositingAllEthBalance'
+  | 'generateAmountExceedsDaiYieldFromDepositingCollateral'
+  | 'generateAmountExceedsDaiYieldFromDepositingCollateralAtNextPrice'
+  | 'generateAmountExceedsDebtCeiling'
+  | 'generateAmountLessThanDebtFloor'
+  | 'generateAmountMoreThanMaxFlashAmount'
+  | 'customAllowanceAmountExceedsMaxUint256'
+  | 'customAllowanceAmountLessThanDepositAmount'
+  | 'ledgerWalletContractDataDisabled'
+  | 'exchangeError'
+  | 'withdrawAmountExceedsFreeCollateral'
+  | 'withdrawAmountExceedsFreeCollateralAtNextPrice'
+  | 'generateAmountExceedsDaiYieldFromTotalCollateral'
+  | 'generateAmountExceedsDaiYieldFromTotalCollateralAtNextPrice'
+  | 'paybackAmountExceedsDaiBalance'
+  | 'paybackAmountExceedsVaultDebt'
+  | 'debtWillBeLessThanDebtFloor'
+  | 'customCollateralAllowanceAmountExceedsMaxUint256'
+  | 'customCollateralAllowanceAmountLessThanDepositAmount'
+  | 'customDaiAllowanceAmountExceedsMaxUint256'
+  | 'customDaiAllowanceAmountLessThanPaybackAmount'
+  | 'withdrawCollateralOnVaultUnderDebtFloor'
+  | 'shouldShowExchangeError'
+  | 'hasToDepositCollateralOnEmptyVault'
+  | 'depositCollateralOnVaultUnderDebtFloor'
 
 interface ErrorMessagesHandler {
   generateAmountLessThanDebtFloor?: boolean
