+++ features/openGuniVault/guniOpenMultiplyVaultValidations.ts
@@ -1,9 +1,5 @@
-import { errorMessagesHandler } from '../form/errorMessagesHandler'
-import { warningMessagesHandler } from '../form/warningMessagesHandler'
-import {
-  VaultErrorMessage,
-  VaultWarningMessage,
-} from '../openMultiplyVault/openMultiplyVaultValidations'
+import { errorMessagesHandler, VaultErrorMessage } from '../form/errorMessagesHandler'
+import { VaultWarningMessage, warningMessagesHandler } from '../form/warningMessagesHandler'
 import { OpenGuniVaultState } from './openGuniVault'
 
 export function validateGuniErrors(state: OpenGuniVaultState): OpenGuniVaultState {
