+++ features/openMultiplyVault/openMultiplyVault.ts
@@ -14,6 +14,8 @@ import { curry } from 'lodash'
 import { combineLatest, iif, merge, Observable, of, Subject, throwError } from 'rxjs'
 import { first, map, scan, shareReplay, switchMap, tap } from 'rxjs/operators'
 
+import { VaultErrorMessage } from '../form/errorMessagesHandler'
+import { VaultWarningMessage } from '../form/warningMessagesHandler'
 import {
   applyExchange,
   createExchangeChange$,
@@ -50,12 +52,7 @@ import {
   setAllowance,
 } from './openMultiplyVaultTransactions'
 import { applyOpenVaultTransition, OpenVaultTransitionChange } from './openMultiplyVaultTransitions'
-import {
-  validateErrors,
-  validateWarnings,
-  VaultErrorMessage,
-  VaultWarningMessage,
-} from './openMultiplyVaultValidations'
+import { validateErrors, validateWarnings } from './openMultiplyVaultValidations'
 
 interface OpenVaultInjectedOverrideChange {
   kind: 'injectStateOverride'
